Disable AST info logs and add fs testMake module manager and modules public
Log an error if the current module isn't found in ModuleManager when
exporting a public item.Make modules and module_manager public

Register the main script as a "main" module in the ModuleManager when
evaluating a script, and insert it into the manager. Add a runtime error
log when attempting to export from a module that isn't present.